backend/requirements.txt
numpy
langchain-core 
langchain-community 
langchain-nvidia-ai-endpoints 
fastapi 
gradio 
langserve 
pyngrok 
faiss-cpu 
sse-starlette 
nest-asyncio 
uvicorn 
pypdf 
langchain-openai
python-dotenv
mysql-connector-python
boto3
